requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cryptography==45.0.6
cymem==2.0.11
Django==5.2.5
docx2txt==0.9
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.19.1
fsspec==2025.7.0
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lxml==6.0.0
marisa-trie==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
numpy==2.3.2
packaging==25.0
pdfminer.six==20250506
preshed==3.0.10
protobuf==6.32.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
python-docx==1.2.0
RapidFuzz==3.13.0
requests==2.32.5
rich==14.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sqlparse==0.5.3
srsly==2.5.1
striprtf==0.0.29
sympy==1.14.0
thinc==8.3.6
torch==2.8.0
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
